+++ NEWS.md
@@ -1,6 +1,6 @@
 # goshawk 0.1.5.9000
 
-* LLOQ/ULOQ explanations no longer shown when loq_legend = FALSE in g_boxplot.R and g_scatterplot.R
+* LLOQ/ULOQ footnotes no longer shown when loq_legend = FALSE in g_boxplot.R and g_scatterplot.R
 * issues with no news:
 
 # goshawk 0.1.5